Unified method for creating packets in APIPacketPicker
From now on, the only method for picking atmotube sensor's packets
is called 'pick_atmotube_api_packets_from_last_timestamp_on()'.

This method takes the last sensor timestamp and use it to understand
if there are new packets to be inserted into the database.

The timestamp may be an EMPTY_STRING. In this case it is ignored.